+++ src/ccip/CCIPLocalSimulator.sol
@@ -4,7 +4,7 @@ pragma solidity ^0.8.19;
 import {WETH9} from "../shared/WETH9.sol";
 import {LinkToken} from "../shared/LinkToken.sol";
 import {BurnMintERC677Helper} from "./BurnMintERC677Helper.sol";
-import {MockCCIPRouter} from "@chainlink/contracts-ccip/contracts/test/mocks/MockRouter.sol";
+import {MockCCIPRouter} from "../vendor/chainlink-ccip/test/mocks/MockRouter.sol";
 import {IRouterClient} from "@chainlink/contracts-ccip/contracts/interfaces/IRouterClient.sol";
 import {IERC20} from "../vendor/openzeppelin-solidity/v4.8.3/contracts/token/ERC20/IERC20.sol";
 import {SafeERC20} from "../vendor/openzeppelin-solidity/v4.8.3/contracts/token/ERC20/utils/SafeERC20.sol";